[CHORE] selectively merging in changes from OgreKit 3.0.2 Merging in from: https://github.com/sonoisa/OgreKit/commit/7f7ab1e452336fb73f23fcc57451e61919f87b3d
Only the regular expression subfolder, because the UI versions are unused by us.

Note that we kept our use of @synchronized in `OGRegularExpression` as this was at least needed in the past.

+++ OgreKit/RegularExpression/NSString_OgreKitAdditions.m
@@ -3,8 +3,8 @@
  * Project: OgreKit
  *
  * Creation Date: Feb 29 2004
- * Author: Isao Sonobe <sonoisa (AT) muse (DOT) ocn (DOT) ne (DOT) jp>
- * Copyright: Copyright (c) 2004 Isao Sonobe, All rights reserved.
+ * Author: Isao Sonobe <sonoisa@gmail.com>
+ * Copyright: Copyright (c) 2004-2018 Isao Sonobe, All rights reserved.
  * License: OgreKit License
  *
  * Encoding: UTF8
@@ -76,14 +76,14 @@
 /***********
  * Replace *
  ***********/
-- (unsigned)replaceOccurrencesOfRegularExpressionString:(NSString*)expressionString 
+- (NSUInteger)replaceOccurrencesOfRegularExpressionString:(NSString*)expressionString
 	withString:(NSString*)replaceString 
 	options:(unsigned)options 
 	range:(NSRange)searchRange
 {
 	OGRegularExpression *regex = [OGRegularExpression regularExpressionWithString:expressionString
 		options:options];
-	unsigned	numberOfReplacement = 0;
+	NSUInteger	numberOfReplacement = 0;
 	NSString	*replacedString = [regex replaceString:self 
 		withString:replaceString 
 		options:options
